design receive page
