--- src/components/Circles/styles.module.scss
@@ -1,27 +1,0 @@
-.circles {
-    position: absolute;
-    bottom: -1em;
-    left: 50%;
-    width: 1140px;
-    margin-left: -570px;
-    height: 700px;
-    transform-origin: 570px 570px;
-    animation: grow 1.5s ease-out;
-
-    .circle {
-        stroke-width: 0.4;
-        fill: rgba(0, 0, 0, 0);
-    }
-}
-
-@keyframes grow {
-    50% {
-        transform: scale(1.5, 1.5);
-        opacity: 0.2;
-    }
-
-    50% {
-        transform: scale(0.5, 0.5);
-        opacity: 0.2;
-    }
-}
