--- src/components/Footer/index.tsx
@@ -1,11 +1,0 @@
-import styles from "./styles.module.scss"
-
-export function Footer() {
-    return (
-        <footer className={styles.footer}>
-            <p>
-                Made by <a href="http://github.com/marvinkome">Marvin Kome</a>
-            </p>
-        </footer>
-    )
-}
