--- src/components/Footer/styles.module.scss
@@ -1,10 +1,0 @@
-.footer {
-    position: absolute;
-    display: flex;
-    bottom: 0;
-    padding: 0em 25px;
-    z-index: 10;
-    height: 50px;
-    width: 100%;
-    background-color: rgba(231, 235, 241, 0.6);
-}
