--- src/components/Header/styles.module.scss
@@ -1,15 +1,0 @@
-.header {
-    position: absolute;
-    z-index: 10;
-    height: 70px;
-    width: 100%;
-    padding: 10px 25px;
-
-    .logoArea {
-        h1 {
-            margin: 0;
-            margin-top: 15px;
-            line-height: 0.5;
-        }
-    }
-}
